requirements.txt
spotipy
python-dotenv
schedule
python-telegram-bot